backend/requirements.txt
flask
flask-cors
requests
python-dotenv>=1.0.0